Angularjs and Cucumberjs integration
Creates angular "describe" blocks for features and scenarios and "it"
blocks for steps.  Also, now always tells cucumber that the step passed
(rather than allowing the client to do so), as we hand off execution of
the steps to Angular. Create an Angular "describe" block for
each Cucumber feature.

Global var check is now added in the source to fix unit test.
Looks for the following global functions:
- featureStart()
- scenarioStart()
- stepStart()

Added support for "And".